+++ ErpNet.FP.Core/Drivers/BgZfpFiscalPrinter.Commands.cs
@@ -2,6 +2,7 @@
 {
     using System;
     using System.Globalization;
+    using System.Numerics;
     using System.Text;
 
     public partial class BgZfpFiscalPrinter : BgFiscalPrinter
@@ -433,8 +434,8 @@
 
             try
             {
-                deviceStatus.Start = int.Parse(fields[0]);
-                deviceStatus.End = int.Parse(fields[1]);
+                deviceStatus.Start = BigInteger.Parse(fields[0]);
+                deviceStatus.End = BigInteger.Parse(fields[1]);
             }
             catch (Exception e)
             {
@@ -473,7 +474,7 @@
             return Request(CommandCloseNonFiscalReceipt);
         }
 
-        public virtual (int?, DeviceStatus) GetCurrentInvoiceNumber()
+        public virtual (BigInteger?, DeviceStatus) GetCurrentInvoiceNumber()
         {
             var invoiceRangeResult = GetInvoiceRange();
